requirements.txt
streamlit==1.26.0
langchain==0.0.267
langchain-google-genai==0.0.5
google-generativeai==0.3.1
python-dotenv==1.0.0